requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
faiss-cpu
pypdf
python-dotenv